backend/requirements.txt
fastapi
uvicorn
openai
numpy
python-dotenv
tqdm
faiss-cpu
